extract DefaultPreviews
