+++ app/src/main/java/ly/david/mbjc/ui/artist/details/ArtistDetailsScreen.kt
@@ -14,7 +14,7 @@ import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.TextWithHeadingRes
 import ly.david.mbjc.ui.common.listitem.InformationListSeparatorHeader
 import ly.david.mbjc.ui.common.listitem.LifeSpanText
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 
 @Composable
