--- app/src/main/java/ly/david/mbjc/ui/common/preview/DefaultPreviews.kt
@@ -1,8 +1,0 @@
-package ly.david.mbjc.ui.common.preview
-
-import android.content.res.Configuration
-import androidx.compose.ui.tooling.preview.Preview
-
-@Preview(name = "Light")
-@Preview(name = "Dark", uiMode = Configuration.UI_MODE_NIGHT_YES)
-annotation class DefaultPreviews
