+++ app/src/main/java/ly/david/mbjc/ui/common/topappbar/ScrollableTopAppBar.kt
@@ -38,7 +38,7 @@ import kotlinx.coroutines.delay
 import ly.david.data.network.MusicBrainzResource
 import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.ResourceIcon
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 
 /**
