+++ app/src/main/java/ly/david/mbjc/ui/event/EventListItem.kt
@@ -18,7 +18,7 @@ import ly.david.data.domain.EventListItemModel
 import ly.david.data.getLifeSpanForDisplay
 import ly.david.mbjc.ExcludeFromJacocoGeneratedReport
 import ly.david.mbjc.ui.common.listitem.DisambiguationText
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 import ly.david.ui.common.theme.TextStyles
