+++ app/src/main/java/ly/david/mbjc/ui/history/HistoryListItem.kt
@@ -22,7 +22,7 @@ import ly.david.mbjc.ui.common.ResourceIcon
 import ly.david.mbjc.ui.common.SMALL_COVER_ART_SIZE
 import ly.david.mbjc.ui.common.getDisplayTextRes
 import ly.david.mbjc.ui.common.listitem.SwipeToDeleteListItem
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 import ly.david.ui.common.theme.TextStyles
