+++ app/src/main/java/ly/david/mbjc/ui/release/details/ReleaseDetailsScreen.kt
@@ -27,7 +27,7 @@ import ly.david.mbjc.ui.common.TextWithHeadingRes
 import ly.david.mbjc.ui.common.coverart.BigCoverArt
 import ly.david.mbjc.ui.common.listitem.InformationListSeparatorHeader
 import ly.david.mbjc.ui.common.listitem.ListSeparatorHeader
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.mbjc.ui.label.LabelListItem
 import ly.david.ui.common.theme.PreviewTheme
