+++ app/src/main/java/ly/david/mbjc/ui/settings/components/SettingWithDialogChoices.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.unit.dp
 import ly.david.mbjc.ExcludeFromJacocoGeneratedReport
 import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.dialog.MultipleChoiceDialog
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.mbjc.ui.settings.AppPreferences
 import ly.david.ui.common.theme.PreviewTheme
 import ly.david.ui.common.theme.TextStyles
