+++ app/src/main/java/ly/david/mbjc/ui/stats/AddReleaseGroupsSection.kt
@@ -21,7 +21,7 @@ import ly.david.data.getDisplayTypes
 import ly.david.data.persistence.releasegroup.ReleaseGroupTypeCount
 import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.listitem.ListSeparatorHeader
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 import ly.david.ui.common.theme.TextStyles
