+++ app/src/main/java/ly/david/mbjc/ui/stats/AddReleasesSection.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.listitem.ListSeparatorHeader
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 
 internal fun LazyListScope.addResourcesStatsSection(
