+++ app/src/main/java/ly/david/mbjc/ui/stats/StatsScreen.kt
@@ -9,7 +9,7 @@ import ly.david.data.persistence.relation.RelationTypeCount
 import ly.david.data.persistence.releasegroup.ReleaseGroupTypeCount
 import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.Tab
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 
 /**
