+++ app/src/main/java/ly/david/mbjc/ui/work/WorkListItem.kt
@@ -12,7 +12,7 @@ import ly.david.data.common.ifNotNullOrEmpty
 import ly.david.data.domain.WorkListItemModel
 import ly.david.mbjc.ExcludeFromJacocoGeneratedReport
 import ly.david.mbjc.ui.common.listitem.DisambiguationText
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 import ly.david.ui.common.theme.TextStyles
