+++ app/src/main/java/ly/david/mbjc/ui/work/details/WorkDetailsScreen.kt
@@ -16,7 +16,7 @@ import ly.david.mbjc.ui.common.TextWithHeading
 import ly.david.mbjc.ui.common.TextWithHeadingRes
 import ly.david.mbjc.ui.common.listitem.AttributesListSeparatorHeader
 import ly.david.mbjc.ui.common.listitem.InformationListSeparatorHeader
-import ly.david.mbjc.ui.common.preview.DefaultPreviews
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 
 @Composable
